requirements.txt
annotated-types==0.6.0
blinker==1.8.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.0
pytest==8.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
